% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChildAppSession.R
\name{createChildAppSession}
\alias{createChildAppSession}
\title{appSession for multi node process, this creates a single appSession}
\usage{
createChildAppSession(selectNames, transcriptomes = allTranscriptomes,
  outputPath = outputPath, scratchPath, fastqPath = fastqPath,
  fastaPath = fastaPath, bootstrapValue, kmer = kmer,
  biasPseudobam = biasPseudobam, fileJSON = fileJSON)
}
\arguments{
\item{selectNames}{vector of name item content parsed from json}

\item{transcriptomes}{vector of transcriptomes}

\item{outputPath}{string path to output}

\item{scratchPath}{string path to scratch directory used}

\item{fastqPath}{string path to fastqPath}

\item{bootstrapValue}{string numeric value parsed by json}

\item{kmer}{numeric kmer size used to reference}

\item{fileJSON}{file json from basespace}

\item{bias}{boolean for bias correction during quant}

\item{pseudobam}{boolean for pseudobam creation}
}
\value{
appSession    list of appSession for quantification
}
\description{
appSession for multi node process, this creates a single appSession
}

