% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChildOutputDirectory.R
\name{createChildOutputDirectory}
\alias{createChildOutputDirectory}
\title{create an ouptut Directory with correct paths for uploading to basespace}
\usage{
createChildOutputDirectory(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{list of input ids created by formbuilder used to parse json file}

\item{fileJSON}{file json read from appSession}
}
\value{
outputPath   a path to output directory
}
\description{
create an ouptut Directory with correct paths for uploading to basespace
}

