% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChildReference.R
\name{createChildReference}
\alias{createChildReference}
\title{create a reference and copy files to scratch directory}
\usage{
createChildReference(selectNames, fileJSON, transcriptomes, fastaPath,
  scratchFastaPath)
}
\arguments{
\item{selectNames}{vector of list of items from JSON}

\item{fileJSON}{list of json elements parsed by jsonlite}

\item{transcriptomes}{vector of input transcriptomes default or manual}

\item{fastaPath}{string path to fasta path}

\item{scratchFastaPath}{string to path to scratch directory}
}
\value{
transcriptomes list of transcriptomes input for quantification
}
\description{
create a reference and copy files to scratch directory
}

