% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDesignMatrix.R
\name{createDesignMatrix}
\alias{createDesignMatrix}
\title{create the Design matrix from comparison and control inputs on basespace}
\usage{
createDesignMatrix(selectNames)
}
\arguments{
\item{selectNames}{vector of input ids from baseSpace input form}
}
\value{
design      matrix of a one group comparison default
}
\description{
create the Design matrix from comparison and control inputs on basespace
}

