% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBootStrapValue.R
\name{findBootStrapValue}
\alias{findBootStrapValue}
\title{finds the bootstrap value input by user on the basespace app}
\usage{
findBootStrapValue(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{vector of strings name ids input by user}

\item{fileJSON}{json file parsed from basespace}
}
\value{
bootstrapValue  string numeric for bootstrap
}
\description{
finds the bootstrap value input by user on the basespace app
}

