% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChildTranscriptInput.R
\name{findChildTranscriptInput}
\alias{findChildTranscriptInput}
\title{finds the transcripts input by user for creating a transcriptome creates a hash table by storing things into the environment}
\usage{
findChildTranscriptInput(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{vector of input id's used for parsing the json file}

\item{fileJSON}{json file from basespace}
}
\value{
transcriptomes   vector of transcriptomes
}
\description{
finds the transcripts input by user for creating a transcriptome creates a hash table by storing things into the environment
}

