% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findreadendType.R
\name{findreadendType}
\alias{findreadendType}
\title{find the read type single end or paired end input from the basespace interface or imported from SRA, for single end reads the length mean is found from find the average length of the sequences}
\usage{
findreadendType(selectNames, fileJSON)
}
\arguments{
\item{selectNames}{vector of input ids from basespace form}

\item{fileJSON}{json file from basespace}
}
\value{
endType   a flag singleEnd,with length stats, or paireEnd passed into appSession
}
\description{
find the read type single end or paired end input from the basespace interface or imported from SRA, for single end reads the length mean is found from find the average length of the sequences
}

