% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutputReports.R
\name{generateOutputReports}
\alias{generateOutputReports}
\title{this runs generateOutputReports.Rmd triggered by a shell script to report the kallisto version used for quantifying transcripts}
\usage{
generateOutputReports(outputDir = appSession$outputPath,
  inputDir = "/outputReport/")
}
\arguments{
\item{outputDir}{output directory to put the html file which should be the app Session output path}

\item{inputDir}{this is the input directory which contains the Rmd file for rendering}
}
\value{
return an html
}
\description{
this runs generateOutputReports.Rmd triggered by a shell script to report the kallisto version used for quantifying transcripts
}

