% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processChildNode.R
\name{processChildNode}
\alias{processChildNode}
\title{runs the multi-node template on baseSpace parses each child node appSession.JSON. the /data/input/appSesson is explicitly defined in the callback.js.  this script is to organize the running of child nodes; designed to process the kallisto quant and analysis on single node.}
\usage{
processChildNode()
}
\description{
runs the multi-node template on baseSpace parses each child node appSession.JSON. the /data/input/appSesson is explicitly defined in the callback.js.  this script is to organize the running of child nodes; designed to process the kallisto quant and analysis on single node.
}

