% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameChildAppSession.R
\name{renameChildAppSessionID}
\alias{renameChildAppSessionID}
\title{rename the samples from session ID to sample Name for multi-node template}
\usage{
renameChildAppSessionID(selectNames, appSession, fileJSON)
}
\arguments{
\item{selectNames}{list of input ids created by form builder on basespace, used to parse the json file}

\item{appSession}{list of appSession flags used for quantification}

\item{fileJSON}{json file from basespace}
}
\value{
DF             data frame of appSession ID and sample Names
}
\description{
rename the samples from session ID to sample Name for multi-node template
}

